Update LambdaCustomResource to SDK V3.
There is a lot of code for AwsCustomResource with SDK V3 that is not published
to NPM. To get around that, we simply copy the CDK code for AwsCustomResource into
custom-resources/handlers/private/from_cdk so we can use that code without having
to 'own' it (in terms of testing etc).

Note we should periodically run the update script here to get any updates from the
CDK custom resource handler code.