feat:  fix findIncluded and LambdaCustomResource properties. (#12)
* fix:  findIncluded with multiple CfnIncludes.

Found bug, wrote failing test, fixed bug, test passes.

* fix:  Move IConstructFactory to Singleton.

Only place the interface is used, it should live in that file.

* feat:  ITask extends ITrigger.

ITrigger interface is simple, useful, and related to ITask.
Made sense to extend.  Hopefully we won't regret this in the future.

* feat:  Implement more of AwsCustomResourceProps.

Only one left is installLatestAwsSdk.

* feat:  Implement most of AwsCustomResourceProps.

A few needed to be passed to the handler function, and that
is now done.  Also needed to implement props.removalPolicy, and that
is also done.  What is left is installLatestAwsSdk.