fix:  Fix TransformHost.of.
TransformHost.of wasn't correctly getting the TransformHost of constructs
in all cases.  It needs to just be a ConstructService with a factory, not
a ConstructTreeService, which was giving wrong answers.

Also, TransformHost.of and TransformHost.isTransformHost need to call
ensureHosted to ensure that antecedent CfnElements and Stacks will always
be recognized as TransformHosts.

All this makes finding Orders work better, which is the point.